+++ croffle/src/main/java/com/be/croffle/member/MemberJpaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface MemberRepository extends JpaRepository<Member, Long> {
+public interface MemberJpaRepository extends JpaRepository<Member, Long> {
     Optional<Member> findByGoogleId(String googleId);
 }
